fix ball detection box sort logic
Fix old ball box sort method, it was done with code:

for box in results.boxes:
    result = box.xyxy.tolist()[0]
    ball_dict[1] = result

which MAKES NO SENCE! YOLO place boxes in order according to descending conf, and this code take the LOWEST CONF BOX!!!!!!

In new logic, it will only treat boxes near by last frame ball box as cur frame ball box. If the box not moving in 30 frame, the area by the box would be banned as it must be not moving thing like small white logo.